Move WeightInfo, Config, Event, LockableCurrencyExt, AccountData into Balances pallet.
