+++ pallets/treasury/src/lib.rs
@@ -45,8 +45,8 @@ use frame_support::{
     weights::Weight,
 };
 use frame_system::ensure_root;
+use pallet_balances::Config as BalancesConfig;
 use pallet_identity as identity;
-use polymesh_common_utilities::traits::balances::Config as BalancesConfig;
 use polymesh_primitives::{constants::TREASURY_PALLET_ID, Beneficiary, IdentityId, GC_DID};
 use sp_runtime::traits::{AccountIdConversion, Saturating};
 use sp_std::prelude::*;
