+++ pallets/utility/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 # Out Pallets
 pallet-balances = { workspace = true, default-features = false }
 pallet-permissions = { workspace = true, default-features = false }
-polymesh-common-utilities = { workspace = true, default-features = false }
 pallet-identity = { workspace = true, default-features = false }
 polymesh-primitives = { workspace = true, default-features = false }
 
@@ -35,7 +34,6 @@ std = [
 	"sp-io/std",
 	"sp-runtime/std",
 	"sp-std/std",
-	"polymesh-common-utilities/std",
 	"polymesh-primitives/std",
 	"pallet-balances/std",
 	"pallet-identity/std",
@@ -45,6 +43,5 @@ runtime-benchmarks = [
 		"frame-benchmarking/runtime-benchmarks",
 		"frame-support/runtime-benchmarks",
 		"frame-system/runtime-benchmarks",
-    "polymesh-common-utilities/runtime-benchmarks",
 ]
 try-runtime = ["frame-support/try-runtime"]
