+++ pallets/utility/src/lib.rs
@@ -84,9 +84,9 @@ use sp_runtime::traits::{BadOrigin, Dispatchable};
 use sp_runtime::{traits::Verify, DispatchError, RuntimeDebug};
 use sp_std::prelude::*;
 
+use pallet_balances::Config as BalancesConfig;
 use pallet_identity::{Config as IdentityConfig, Context};
 use pallet_permissions::with_call_metadata;
-use polymesh_common_utilities::balances::Config as BalancesConfig;
 use polymesh_primitives::{identity::AuthorizationNonce, traits::CheckCdd, IdentityId};
 
 type Identity<T> = pallet_identity::Module<T>;
